requirements.txt
flask
azure-cosmos
python-dotenv
azure-ai-documentintelligence
azure-identity
pydantic
langchain
langchain_core
langchain_community
langchain_openai
tiktoken
requests
fuzzywuzzy
python-Levenshtein
azure-storage-blob
azure-servicebus